This implements in C++ the basic operations for arbitrarily sized tensors: contraction, convolution, outer product, partial outer product, (and multiplication by matrix as a special case of contraction). Also provides the ability to statically switch between lexicographically ordering and colexicographically ordering the flat form of the tensor. More testing is necessary but all of the operations appeared to work in several small cases.
This was just a first attempt to learn the operations and little consideration has been put into
optimization. There is some basic code to help with adhoc profiling. I think we should maintain
C++ and CUDA versions of these methods to get an idea about the actual speedup
of using CUDA.

Builds on my mac and on my UMIACS account (with the latest gcc module added).
Call   make && ./bin/OPS