Fixes two errors. The first is that when the user does a convolution, but does not save the convolution as an output index, then I was forgetting to finally contract out the index after the convolution. I think ConvXd can actually do this operation atomically, so a \todo is to implement additional atomic calls to ConvXd which handle this case
The second is I was incorrectly evaluating >= 3 tensors in the left-to-right manner. Consider the left-to-right
evaluation sequence ((((A, B), C, D, E))). If A or B shares an index with C, which is not an output index in the
final tensor, then I was forgetting to store that index in the intermediate object produced by (A, B).